Move all unit tests in new folder
