+++ test/unit/component/parser/json/BpmnJsonParser.parallelGateway.test.ts
@@ -1,22 +1,22 @@
-import { ShapeBpmnElementKind } from '../../../../src/model/bpmn/shape/ShapeBpmnElementKind';
+import { ShapeBpmnElementKind } from '../../../../../src/model/bpmn/shape/ShapeBpmnElementKind';
 import { parseJsonAndExpectOnlyFlowNodes, verifyShape } from './JsonTestUtils';
 
-describe('parse bpmn as json for start event', () => {
-  it('json containing one process with a single start event', () => {
+describe('parse bpmn as json for parallel gateway', () => {
+  it('json containing one process with a single parallel gateway', () => {
     const json = `{
                 "definitions" : {
                     "process": {
-                        "startEvent": {
-                            "id":"event_id_0",
-                            "name":"event name"
+                        "parallelGateway": {
+                            "id":"parallelGateway_id_0",
+                            "name":"parallelGateway name"
                         }
                     },
                     "BPMNDiagram": {
                         "name":"process 0",
                         "BPMNPlane": {
                             "BPMNShape": {
-                                "id":"shape_startEvent_id_0",
-                                "bpmnElement":"event_id_0",
+                                "id":"shape_parallelGateway_id_0",
+                                "bpmnElement":"parallelGateway_id_0",
                                 "Bounds": { "x": 362, "y": 232, "width": 36, "height": 45 }
                             }
                         }
@@ -27,10 +27,10 @@ describe('parse bpmn as json for start event', () => {
     const model = parseJsonAndExpectOnlyFlowNodes(json, 1);
 
     verifyShape(model.flowNodes[0], {
-      shapeId: 'shape_startEvent_id_0',
-      bpmnElementId: 'event_id_0',
-      bpmnElementName: 'event name',
-      bpmnElementKind: ShapeBpmnElementKind.EVENT_START,
+      shapeId: 'shape_parallelGateway_id_0',
+      bpmnElementId: 'parallelGateway_id_0',
+      bpmnElementName: 'parallelGateway name',
+      bpmnElementKind: ShapeBpmnElementKind.GATEWAY_PARALLEL,
       boundsX: 362,
       boundsY: 232,
       boundsWidth: 36,
@@ -38,14 +38,14 @@ describe('parse bpmn as json for start event', () => {
     });
   });
 
-  it('json containing one process declared as array with a single start event', () => {
+  it('json containing one process declared as array with a single parallel gateway', () => {
     const json = `{
                 "definitions": {
                     "process": [
                         {
-                            "startEvent": {
-                                "id":"event_id_1",
-                                "name":"event name"
+                            "parallelGateway": {
+                                "id":"parallelGateway_id_1",
+                                "name":"parallelGateway name"
                             }
                         }
                     ],
@@ -53,8 +53,8 @@ describe('parse bpmn as json for start event', () => {
                         "name":"process 0",
                         "BPMNPlane": {
                             "BPMNShape": {
-                                "id":"shape_startEvent_id_1",
-                                "bpmnElement":"event_id_1",
+                                "id":"shape_parallelGateway_id_1",
+                                "bpmnElement":"parallelGateway_id_1",
                                 "Bounds": { "x": 362, "y": 232, "width": 36, "height": 45 }
                             }
                         }
@@ -65,10 +65,10 @@ describe('parse bpmn as json for start event', () => {
     const model = parseJsonAndExpectOnlyFlowNodes(json, 1);
 
     verifyShape(model.flowNodes[0], {
-      shapeId: 'shape_startEvent_id_1',
-      bpmnElementId: 'event_id_1',
-      bpmnElementName: 'event name',
-      bpmnElementKind: ShapeBpmnElementKind.EVENT_START,
+      shapeId: 'shape_parallelGateway_id_1',
+      bpmnElementId: 'parallelGateway_id_1',
+      bpmnElementName: 'parallelGateway name',
+      bpmnElementKind: ShapeBpmnElementKind.GATEWAY_PARALLEL,
       boundsX: 362,
       boundsY: 232,
       boundsWidth: 36,
@@ -76,16 +76,16 @@ describe('parse bpmn as json for start event', () => {
     });
   });
 
-  it('json containing one process with an array of start events with name & without name', () => {
+  it('json containing one process with an array of parallel gateways with name & without name', () => {
     const json = `{
                 "definitions" : {
                     "process": {
-                        "startEvent": [
+                        "parallelGateway": [
                           {
-                              "id":"event_id_0",
-                              "name":"event name"
+                              "id":"parallelGateway_id_0",
+                              "name":"parallelGateway name"
                           }, {
-                              "id":"event_id_1"
+                              "id":"parallelGateway_id_1"
                           }
                         ]
                     },
@@ -94,12 +94,12 @@ describe('parse bpmn as json for start event', () => {
                         "BPMNPlane": {
                             "BPMNShape": [
                               {
-                                "id":"shape_startEvent_id_0",
-                                "bpmnElement":"event_id_0",
+                                "id":"shape_parallelGateway_id_0",
+                                "bpmnElement":"parallelGateway_id_0",
                                 "Bounds": { "x": 362, "y": 232, "width": 36, "height": 45 }
                               }, {
-                                "id":"shape_startEvent_id_1",
-                                "bpmnElement":"event_id_1",
+                                "id":"shape_parallelGateway_id_1",
+                                "bpmnElement":"parallelGateway_id_1",
                                 "Bounds": { "x": 365, "y": 235, "width": 35, "height": 46 }
                               }
                             ]
@@ -111,20 +111,20 @@ describe('parse bpmn as json for start event', () => {
     const model = parseJsonAndExpectOnlyFlowNodes(json, 2);
 
     verifyShape(model.flowNodes[0], {
-      shapeId: 'shape_startEvent_id_0',
-      bpmnElementId: 'event_id_0',
-      bpmnElementName: 'event name',
-      bpmnElementKind: ShapeBpmnElementKind.EVENT_START,
+      shapeId: 'shape_parallelGateway_id_0',
+      bpmnElementId: 'parallelGateway_id_0',
+      bpmnElementName: 'parallelGateway name',
+      bpmnElementKind: ShapeBpmnElementKind.GATEWAY_PARALLEL,
       boundsX: 362,
       boundsY: 232,
       boundsWidth: 36,
       boundsHeight: 45,
     });
     verifyShape(model.flowNodes[1], {
-      shapeId: 'shape_startEvent_id_1',
-      bpmnElementId: 'event_id_1',
+      shapeId: 'shape_parallelGateway_id_1',
+      bpmnElementId: 'parallelGateway_id_1',
       bpmnElementName: undefined,
-      bpmnElementKind: ShapeBpmnElementKind.EVENT_START,
+      bpmnElementKind: ShapeBpmnElementKind.GATEWAY_PARALLEL,
       boundsX: 365,
       boundsY: 235,
       boundsWidth: 35,
