+++ test/unit/component/parser/json/BpmnJsonParser.sequenceFlow.test.ts
