+++ test/unit/component/parser/xml/BpmnXmlParser.process.test.ts
@@ -1,4 +1,4 @@
-import BpmnXmlParser from '../../../../src/component/parser/xml/BpmnXmlParser';
+import BpmnXmlParser from '../../../../../src/component/parser/xml/BpmnXmlParser';
 import { verifyDefinitions, verifyProperties } from './XMLTestUtils';
 
 describe('parse bpmn as xml for process', () => {
