+++ test/unit/component/parser/xml/BpmnXmlParser.sequenceFlow.test.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import BpmnXmlParser from '../../../../src/component/parser/xml/BpmnXmlParser';
+import BpmnXmlParser from '../../../../../src/component/parser/xml/BpmnXmlParser';
 import { verifyDefinitions, verifyIsNotEmptyArray, verifyProperties } from './XMLTestUtils';
 
 describe('parse bpmn as xml for sequence flow', () => {
