+++ test/unit/component/parser/xml/XMLTestUtils.ts
